Fix pool update job
The pool update job was not launched because it had the same name.
As Kubernetes requires names be unique across all API versions of the same
resource, the jobs were only triggered once.

This patch adds the jobs a unix timestamp suffix to overcome that
problem.

It also sorts the pool resources to create a comparable hash between
jobs.